start case finished

+++ src/main/java/rt/bot/cases/StartCase.java
@@ -18,6 +18,7 @@ import rt.bot.telegram.constants.Menu;
 import rt.bot.telegram.constants.Text;
 import rt.bot.telegram.handlers.BotAdminHandler;
 import rt.bot.telegram.handlers.TariffAvailabilityHandler;
+import rt.bot.telegram.handlers.VipChannelHandler;
 
 @Slf4j
 @Component
@@ -29,6 +30,7 @@ public class StartCase implements Case {
     private final MessageRemover remover;
     private final BotAdminHandler botAdminHandler;
     private final TariffAvailabilityHandler tariffAvailabilityHandler;
+    private final VipChannelHandler vipChannelHandler;
     private final UserService userService;
     private final ChannelService channelService;
     private final SubscriptionService subscriptionService;
@@ -38,16 +40,22 @@ public class StartCase implements Case {
         Long userId = TelegramUtils.extractUserIdFromUpdate(update);
         String userMsg = TelegramUtils.extractUserTextFromUpdate(update);
 
-        if (isStartTariffRequest(botUser, userMsg)) {
+        if (isStartVipRequest(botUser, userMsg)) {
+            vipChannelHandler.process(botUser, userId, userMsg);
+            return;
+        }
+        if (isCancel(userMsg)) {
+            handleCancelParticipation(botUser, userId);
+        } else if (isStartTariffRequest(botUser, userMsg)) {
             handleStartRequest(botUser, userId, userMsg);
-        } else if (botIsAddedAsAdmin(update, botUser)) {
+        } else if (botAddedAsAdmin(update, botUser)) {
             handleBotAssignment(update, botUser, userId);
+        } else if (botDismissedDuringRegistration(update, botUser)) {
+            handleBotDismissedDuringRegistration(botUser, userId);
         } else if (subscriptionConfirmed(botUser, userMsg)) {
             handleSubscriptionConfirmation(botUser, userId);
-        } else if (isPaymentConfirmation(botUser, userMsg)) {
-            handlePaymentConfirmation(botUser, userId, userMsg);
-        } else if (isCancel(userMsg)) {
-            handleCancelParticipation(botUser, userId);
+        } else if (paymentConfirmed(botUser, userMsg)) {
+            handlePaymentConfirmation(botUser, userId);
         } else {
             informAboutError(botUser, userId);
         }
@@ -84,9 +92,8 @@ public class StartCase implements Case {
         remover.removeMsg(userId, botUser.getLastMessageIdToDelete());
         botAdminHandler.process(update);
         if (channelService.channelExists(update.getMyChatMember().getChat().getId())) {
-            sender.send(userId, Text.CHANNEL_ADDED_EARLY);
             userService.setTariffAndStatus(botUser, Tariff.NONE, BotUser.DialogStatus.NONE);
-            sender.send(userId, Text.LEARN_ANOTHER_TARIFF, Menu.getTariffMenu());
+            sender.send(userId, Text.CHANNEL_ADDED_EARLY, Menu.of(Text.START_AGAIN));
             return;
         }
         channelService.createNewChannel(botUser, update.getMyChatMember().getChat());
@@ -94,11 +101,15 @@ public class StartCase implements Case {
     }
 
     private void sendChannelsToSubscribe(BotUser botUser, Long userId) {
+        if (botUser.getTariff() == Tariff.TARIFF_3) {
+            handleSubscriptionConfirmation(botUser, userId);
+            return;
+        }
         String channelsList = subscriptionService.getChannelListToSubscribe(botUser);
         if (channelsList == null || channelsList.isBlank()) {
             channelService.removeCancelledChannel(userId);
             userService.setTariffAndStatus(botUser, Tariff.NONE, BotUser.DialogStatus.NONE);
-            sender.send(userId, Text.CANCEL_DUE_TO_LACK_OF_CHANNELS);
+            sender.send(userId, Text.CANCEL_DUE_TO_LACK_OF_CHANNELS, Menu.of(Text.START_AGAIN));
             log.info("Нет каналов для подписки для пользователя с id {}", userId);
             return;
         }
@@ -112,18 +123,23 @@ public class StartCase implements Case {
     private void handleSubscriptionConfirmation(BotUser botUser, Long userId) {
         if (botUser.getTariff() == Tariff.TARIFF_1) {
             finishProcess(botUser, userId);
-        } else {
-            //todo payment service
-            String link = "";
-            userService.setDialogStatus(botUser, BotUser.DialogStatus.WAITING_PAYMENT_CONFIRMATION);
-            sender.send(
-                    userId,
-                    Text.PAYMENT_LINK.formatted(link),
-                    Menu.of(Text.PAYMENT_CONFIRM, Text.CANCEL_PARTICIPATION));
+            return;
         }
+        String link = "";
+        userService.setDialogStatus(botUser, BotUser.DialogStatus.WAITING_PAYMENT_CONFIRMATION);
+        if (botUser.getTariff() == Tariff.TARIFF_2) {
+            link = "t2";
+        } else if (botUser.getTariff() == Tariff.TARIFF_3) {
+            link = "t3";
+        }
+        //todo payment service
+        sender.send(
+                userId,
+                Text.PAYMENT_LINK.formatted(link),
+                Menu.of(Text.PAYMENT_CONFIRM, Text.CANCEL_PARTICIPATION));
     }
 
-    private void handlePaymentConfirmation(BotUser botUser, Long userId, String userMsg) {
+    private void handlePaymentConfirmation(BotUser botUser, Long userId) {
         // todo payment service check payment
         finishProcess(botUser, userId);
     }
@@ -131,14 +147,19 @@ public class StartCase implements Case {
     private void handleCancelParticipation(BotUser botUser, Long userId) {
         channelService.removeCancelledChannel(userId);
         userService.setTariffAndStatus(botUser, Tariff.NONE, BotUser.DialogStatus.NONE);
-        sender.send(userId, Text.CANCEL_PARTICIPATION_SUCCESS);
-        sender.send(userId, Text.GREET, Menu.getTariffMenu(), "MarkdownV2");
+        sender.send(userId, Text.CANCEL_PARTICIPATION_SUCCESS, Menu.of(Text.START_AGAIN));
+    }
+
+    private void handleBotDismissedDuringRegistration(BotUser botUser, Long userId) {
+        channelService.removeCancelledChannel(userId);
+        userService.setTariffAndStatus(botUser, Tariff.NONE, BotUser.DialogStatus.NONE);
+        sender.send(userId, Text.FAILED_PARTICIPATION_DUE_TO_BOT_ADMIN_RIGHTS, Menu.of(Text.START_AGAIN));
     }
 
     private void finishProcess(BotUser botUser, Long userId) {
-        userService.setTariffAndStatus(botUser, Tariff.NONE, BotUser.DialogStatus.NONE);
+        userService.setRoleTariffStatus(botUser, BotUser.Role.USER, Tariff.NONE, BotUser.DialogStatus.NONE);
         channelService.setPaidSince(botUser.getUserId());
-        sender.send(userId, Text.SUCCESS, Menu.removeReplyKeyboard());
+        sender.send(userId, Text.SUCCESS, Menu.of(Text.MY_SUBSCRIPTIONS));
     }
 
     private void informAboutError(BotUser botUser, Long userId) {
@@ -152,7 +173,7 @@ public class StartCase implements Case {
         }
     }
 
-    private boolean botIsAddedAsAdmin(Update update, BotUser botUser) {
+    private boolean botAddedAsAdmin(Update update, BotUser botUser) {
         return TelegramUtils.botIsAddedAsAdmin(update) &&
                 botUser.getDialogStatus() == BotUser.DialogStatus.WAITING_ADD_BOT_AS_ADMIN;
     }
@@ -163,17 +184,30 @@ public class StartCase implements Case {
                 userMsg.startsWith(Text.TARIFF_START);
     }
 
+    private boolean isStartVipRequest(BotUser botUser, String userMsg) {
+        if (botUser.getRole() != BotUser.Role.ADMIN) return false;
+        return (botUser.getDialogStatus() == BotUser.DialogStatus.NONE &&
+                botUser.getTariff() == Tariff.NONE &&
+                userMsg.equals(Text.START_VIP) ||
+                botUser.getDialogStatus() == BotUser.DialogStatus.WAITING_CHANNEL_LINK);
+    }
+
     private boolean subscriptionConfirmed(BotUser botUser, String userMsg) {
         return botUser.getDialogStatus() == BotUser.DialogStatus.WAITING_SUBSCRIPTION_CONFIRMATION &&
                 userMsg.equals(Text.SUBSCRIPTION_CONFIRMATION);
     }
 
-    private boolean isPaymentConfirmation(BotUser botUser, String userMsg) {
+    private boolean paymentConfirmed(BotUser botUser, String userMsg) {
         if (botUser.getDialogStatus() != BotUser.DialogStatus.WAITING_PAYMENT_CONFIRMATION) return false;
         return userMsg.equals(Text.PAYMENT_CONFIRM);
     }
 
     private boolean isCancel(String userMsg) {
         return userMsg.equals(Text.CANCEL_PARTICIPATION);
+    }
+
+    private boolean botDismissedDuringRegistration(Update update, BotUser botUser) {
+        return TelegramUtils.botIsDismissed(update) &&
+                botUser.getDialogStatus() != BotUser.DialogStatus.NONE;
     }
 }
